Fix issue seen with icc_rendering_intent.pdf
LCMS2 has an attempted fix in it to cope with "inverted" profiles
whereby it clamps out of range L values (in an Lab space) to 0.
This is a change from previous behaviour, where values < 0 were
clamped to 0, and values > 50 wer clamped to 50.

Having discussed this with Marti (for which many thanks, as ever!)
we've come up with the reformulation here.